+++ scenedetect/cli/context.py
@@ -37,7 +37,7 @@ logger = logging.getLogger('pyscenedetect')
 USER_CONFIG = ConfigRegistry()
 
 
-def parse_timecode(value: Union[str, int, FrameTimecode], frame_rate: float) -> FrameTimecode:
+def parse_timecode(value: str, frame_rate: float, first_index_is_one: bool = False) -> FrameTimecode:
     """Parses a user input string into a FrameTimecode assuming the given framerate.
 
     If value is None, None will be returned instead of processing the value.
@@ -48,6 +48,10 @@ def parse_timecode(value: Union[str, int, FrameTimecode], frame_rate: float) -> 
     if value is None:
         return None
     try:
+        if first_index_is_one and value.isdigit():
+            value = int(value)
+            if value >= 1:
+                value -= 1
         return FrameTimecode(timecode=value, fps=frame_rate)
     except ValueError as ex:
         raise click.BadParameter(
@@ -655,9 +659,9 @@ class CliContext:
         logger.debug('Setting video time:\n    start: %s, duration: %s, end: %s', start, duration,
                      end)
 
-        self.start_time = parse_timecode(start, self.video_stream.frame_rate)
-        self.end_time = parse_timecode(end, self.video_stream.frame_rate)
-        self.duration = parse_timecode(duration, self.video_stream.frame_rate)
+        self.start_time = parse_timecode(start, self.video_stream.frame_rate, first_index_is_one=True)
+        self.end_time = parse_timecode(end, self.video_stream.frame_rate, first_index_is_one=True)
+        self.duration = parse_timecode(duration, self.video_stream.frame_rate, first_index_is_one=True)
         self.time = True
 
         self.options_processed = options_processed_orig